fixed possible overflow InventoryMenu (#46)
* fixed possible overflow InventoryMenu

Now Item::count() returns the same type as Daedalus::GEngineClasses::C_Item.amount

* Revert "fixed possible overflow InventoryMenu"

This reverts commit bac937aae40d04d774ac48d33ac67e7224c6a064.

* fixed possible overflow on InventoryMenu::onTakeStuff()

+++ Game/ui/inventorymenu.cpp
@@ -342,7 +342,7 @@ InventoryMenu::PageLocal &InventoryMenu::activePageSel() {
 #pragma GCC diagnostic ignored "-Wfloat-conversion"
 #endif
 void InventoryMenu::onTakeStuff() { 
-  uint8_t itemCount = 0;
+  uint32_t itemCount = 0;
   auto& page = activePage();
   auto& sel = activePageSel();
   if(sel.sel >= page.size())
@@ -352,7 +352,7 @@ void InventoryMenu::onTakeStuff() {
     ++takeCount;
     itemCount = pow(10,takeCount / 10);
     if(r.count() <= itemCount) {
-      itemCount = r.count();
+      itemCount = uint32_t(r.count());
       takeCount = 0;
     }
   }
